Added some changes for RN composible UI
